+++ benchmarks/markdown_table/scripts/data-update.ts
@@ -1,0 +1,1 @@
+// noop for now, but will be created later.