+++ packages/gatsby-plugin-manifest/src/__tests__/gatsby-ssr.js
@@ -1,12 +1,6 @@
 jest.mock(`fs`, () => {
   return {
-    readFileSync: jest.fn().mockImplementation(() => `someIconImage`),
-  }
-})
-
-jest.mock(`gatsby-core-utils`, () => {
-  return {
-    createContentDigest: jest.fn(() => `contentDigest`),
+    readFileSync: jest.fn().mockImplementation(() => `someIconImage`)
   }
 })
 
@@ -17,7 +11,7 @@ const setHeadComponents = args => (headComponents = headComponents.concat(args))
 
 const ssrArgs = {
   setHeadComponents,
-  pathname: `/`,
+  pathname: `/`
 }
 
 describe(`gatsby-plugin-manifest`, () => {
@@ -32,7 +26,7 @@ describe(`gatsby-plugin-manifest`, () => {
 
     onRenderBody(ssrArgs, {
       icon: true,
-      theme_color: `#000000`,
+      theme_color: `#000000`
     })
 
     headComponents
@@ -50,7 +44,7 @@ describe(`gatsby-plugin-manifest`, () => {
         theme_color: `#000000`,
         include_favicon: false,
         cache_busting_mode: false,
-        legacy: false,
+        legacy: false
       })
       expect(headComponents).toMatchSnapshot()
     })
@@ -61,7 +55,7 @@ describe(`gatsby-plugin-manifest`, () => {
         theme_color_in_head: true,
         include_favicon: false,
         cache_busting_mode: false,
-        legacy: false,
+        legacy: false
       })
       expect(headComponents).toMatchSnapshot()
     })
@@ -72,7 +66,7 @@ describe(`gatsby-plugin-manifest`, () => {
         theme_color_in_head: false,
         include_favicon: false,
         cache_busting_mode: false,
-        legacy: false,
+        legacy: false
       })
       expect(headComponents).toMatchSnapshot()
     })
@@ -82,7 +76,7 @@ describe(`gatsby-plugin-manifest`, () => {
         icon: true,
         include_favicon: false,
         cache_busting_mode: false,
-        legacy: false,
+        legacy: false
       })
       expect(headComponents).toMatchSnapshot()
     })
@@ -90,7 +84,7 @@ describe(`gatsby-plugin-manifest`, () => {
     it(`Adds "icon" and "manifest" links and "theme_color" meta tag to head`, () => {
       onRenderBody(ssrArgs, {
         icon: true,
-        theme_color: `#000000`,
+        theme_color: `#000000`
       })
       expect(headComponents).toMatchSnapshot()
     })
@@ -99,13 +93,13 @@ describe(`gatsby-plugin-manifest`, () => {
       {
         ...ssrArgs,
         pathname: `/about-us`,
-        testName: `Adds correct (default) i18n "manifest" link to head`,
+        testName: `Adds correct (default) i18n "manifest" link to head`
       },
       {
         ...ssrArgs,
         pathname: `/es/sobre-nosotros`,
-        testName: `Adds correct (es) i18n "manifest" link to head`,
-      },
+        testName: `Adds correct (es) i18n "manifest" link to head`
+      }
     ]
 
     i18nArgs.forEach(({ testName, ...args }) =>
@@ -116,13 +110,13 @@ describe(`gatsby-plugin-manifest`, () => {
           localize: [
             {
               start_url: `/de/`,
-              lang: `de`,
+              lang: `de`
             },
             {
               start_url: `/es/`,
-              lang: `es`,
-            },
-          ],
+              lang: `es`
+            }
+          ]
         })
         expect(headComponents).toMatchSnapshot()
       })
@@ -137,7 +131,7 @@ describe(`gatsby-plugin-manifest`, () => {
           theme_color: `#000000`,
           include_favicon: false,
           cache_busting_mode: `none`,
-          theme_color_in_head: false,
+          theme_color_in_head: false
         })
         expect(headComponents).toMatchSnapshot()
       })
@@ -150,17 +144,17 @@ describe(`gatsby-plugin-manifest`, () => {
             {
               src: `/favicons/android-chrome-48x48.png`,
               sizes: `48x48`,
-              type: `image/png`,
+              type: `image/png`
             },
             {
               src: `/favicons/android-chrome-512x512.png`,
               sizes: `512x512`,
-              type: `image/png`,
-            },
+              type: `image/png`
+            }
           ],
           include_favicon: false,
           cache_busting_mode: `none`,
-          theme_color_in_head: false,
+          theme_color_in_head: false
         })
         expect(headComponents).toMatchSnapshot()
       })
@@ -172,14 +166,14 @@ describe(`gatsby-plugin-manifest`, () => {
             {
               src: `/favicons/android-chrome-48x48.png`,
               sizes: `48x48`,
-              type: `image/png`,
+              type: `image/png`
             },
             {
               src: `/favicons/android-chrome-512x512.png`,
               sizes: `512x512`,
-              type: `image/png`,
-            },
-          ],
+              type: `image/png`
+            }
+          ]
         })
         expect(headComponents).toMatchSnapshot()
       })
@@ -191,7 +185,7 @@ describe(`gatsby-plugin-manifest`, () => {
           icon: true,
           legacy: false,
           include_favicon: false,
-          cache_busting_mode: false,
+          cache_busting_mode: false
         })
         expect(headComponents).toMatchSnapshot()
       })
@@ -205,14 +199,14 @@ describe(`gatsby-plugin-manifest`, () => {
             {
               src: `/favicons/android-chrome-48x48.png`,
               sizes: `48x48`,
-              type: `image/png`,
+              type: `image/png`
             },
             {
               src: `/favicons/android-chrome-512x512.png`,
               sizes: `512x512`,
-              type: `image/png`,
-            },
-          ],
+              type: `image/png`
+            }
+          ]
         })
         expect(headComponents).toMatchSnapshot()
       })
@@ -225,16 +219,16 @@ describe(`gatsby-plugin-manifest`, () => {
             {
               src: `/favicons/android-chrome-48x48.png`,
               sizes: `48x48`,
-              type: `image/png`,
+              type: `image/png`
             },
             {
               src: `/favicons/android-chrome-512x512.png`,
               sizes: `512x512`,
-              type: `image/png`,
-            },
+              type: `image/png`
+            }
           ],
           include_favicon: false,
-          cache_busting_mode: false,
+          cache_busting_mode: false
         })
         expect(headComponents).toMatchSnapshot()
       })
@@ -249,15 +243,15 @@ describe(`gatsby-plugin-manifest`, () => {
           {
             src: `/favicons/android-chrome-48x48.png`,
             sizes: `48x48`,
-            type: `image/png`,
+            type: `image/png`
           },
           {
             src: `/favicons/android-chrome-512x512.png`,
             sizes: `512x512`,
-            type: `image/png`,
-          },
+            type: `image/png`
+          }
         ],
-        cache_busting_mode: `name`,
+        cache_busting_mode: `name`
       })
       expect(headComponents).toMatchSnapshot()
     })
@@ -289,7 +283,7 @@ describe(`gatsby-plugin-manifest`, () => {
         icon: true,
         include_favicon: true,
         legacy: false,
-        cache_busting_mode: `none`,
+        cache_busting_mode: `none`
       })
       expect(headComponents).toMatchSnapshot()
     })
@@ -299,7 +293,7 @@ describe(`gatsby-plugin-manifest`, () => {
         icon: true,
         include_favicon: false,
         legacy: false,
-        cache_busting_mode: `none`,
+        cache_busting_mode: `none`
       })
       expect(headComponents).toMatchSnapshot()
     })
@@ -311,16 +305,16 @@ describe(`gatsby-plugin-manifest`, () => {
           {
             src: `/favicons/android-chrome-48x48.png`,
             sizes: `48x48`,
-            type: `image/png`,
+            type: `image/png`
           },
           {
             src: `/favicons/android-chrome-512x512.png`,
             sizes: `512x512`,
-            type: `image/png`,
-          },
+            type: `image/png`
+          }
         ],
         legacy: false,
-        cache_busting_mode: `none`,
+        cache_busting_mode: `none`
       })
       expect(headComponents).toMatchSnapshot()
     })
@@ -332,7 +326,7 @@ describe(`gatsby-plugin-manifest`, () => {
         crossOrigin: `use-credentials`,
         legacy: false,
         include_favicon: false,
-        cache_busting_mode: `none`,
+        cache_busting_mode: `none`
       })
       expect(headComponents).toMatchSnapshot()
     })
@@ -343,7 +337,7 @@ describe(`gatsby-plugin-manifest`, () => {
         crossOrigin: `anonymous`,
         legacy: false,
         include_favicon: false,
-        cache_busting_mode: `none`,
+        cache_busting_mode: `none`
       })
       expect(headComponents).toMatchSnapshot()
     })
@@ -353,7 +347,7 @@ describe(`gatsby-plugin-manifest`, () => {
         icon: true,
         legacy: false,
         include_favicon: false,
-        cache_busting_mode: `none`,
+        cache_busting_mode: `none`
       })
       expect(headComponents).toMatchSnapshot()
     })