+++ packages/gatsby-plugin-mdx/utils/gen-mdx.js
@@ -70,7 +70,7 @@ module.exports = async function genMDX(
     html: undefined,
     scopeImports: [],
     scopeIdentifiers: [],
-    body: undefined,
+    body: undefined
   }
 
   // TODO: a remark and a hast plugin that pull out the ast and store it in results
@@ -116,16 +116,21 @@ export const _frontmatter = ${JSON.stringify(data)}`
       reporter,
       cache,
       pathPrefix,
-      ...helpers,
+      compiler: {
+        parseString: compiler.parse.bind(compiler),
+        generateHTML: ast => mdx(ast, options)
+      },
+      ...helpers
     }
   )
 
   debug(`running mdx`)
   let code = await mdx(content, {
+    filepath: node.fileAbsolutePath,
     ...options,
     remarkPlugins: options.remarkPlugins.concat(
       gatsbyRemarkPluginsAsremarkPlugins
-    ),
+    )
   })
 
   results.rawMDXOutput = `/* @jsx mdx */
@@ -141,7 +146,7 @@ ${code}`
         instance.plugin,
         objRestSpread,
         htmlAttrToJSXAttr,
-        removeExportKeywords,
+        removeExportKeywords
       ],
       presets: [
         require(`@babel/preset-react`),
@@ -150,10 +155,10 @@ ${code}`
           {
             useBuiltIns: `entry`,
             corejs: 2,
-            modules: false,
-          },
-        ],
-      ],
+            modules: false
+          }
+        ]
+      ]
     })
 
     const identifiers = Array.from(instance.state.identifiers)