+++ packages/gatsby-recipes/src/providers/gatsby/utils/constants.js
@@ -1,0 +1,3 @@
+const REQUIRES_KEYS = [`gatsby-source-filesystem`, `gatsby-plugin-page-creator`]
+
+module.exports.REQUIRES_KEYS = REQUIRES_KEYS