+++ packages/gatsby/src/redux/reducers/index.js
@@ -1,5 +1,7 @@
 const reduxNodes = require(`./nodes`)
 const lokiNodes = require(`../../db/loki/nodes`).reducer
+import { redirectsReducer } from "./redirects"
+import { reducer as logReducer } from "gatsby-cli/lib/reporter/redux/reducer"
 
 const backend = process.env.GATSBY_DB_NODES || `redux`
 
@@ -59,12 +61,12 @@ module.exports = {
   jobsV2: require(`./jobsv2`),
   webpack: require(`./webpack`),
   webpackCompilationHash: require(`./webpack-compilation-hash`),
-  redirects: require(`./redirects`),
+  redirects: redirectsReducer,
   babelrc: require(`./babelrc`),
   schemaCustomization: require(`./schema-customization`),
   themes: require(`./themes`),
-  logs: require(`gatsby-cli/lib/reporter/redux/reducer`),
+  logs: logReducer,
   inferenceMetadata: require(`./inference-metadata`),
   pageDataStats: require(`./page-data-stats`),
-  pageData: require(`./page-data`),
+  pageData: require(`./page-data`)
 }