+++ packages/gatsby-plugin-sharp/CHANGELOG.md
@@ -3,6 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.7](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.5.6...gatsby-plugin-sharp@2.5.7) (2020-04-24)
+
+**Note:** Version bump only for package gatsby-plugin-sharp
+
+## [2.5.6](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.5.5...gatsby-plugin-sharp@2.5.6) (2020-04-17)
+
+### Bug Fixes
+
+- **gatsby-plugin-sharp:** error if we can't determine dimensio… ([#23156](https://github.com/gatsbyjs/gatsby/issues/23156)) ([e11c970](https://github.com/gatsbyjs/gatsby/commit/e11c970))
+- wrap ignore pattern in quotes ([#23176](https://github.com/gatsbyjs/gatsby/issues/23176)) ([7563db6](https://github.com/gatsbyjs/gatsby/commit/7563db6))
+
+## [2.5.5](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.5.4...gatsby-plugin-sharp@2.5.5) (2020-04-16)
+
+### Bug Fixes
+
+- **gatsby-plugin-sharp:** trim option failed ([#22997](https://github.com/gatsbyjs/gatsby/issues/22997)) ([0558443](https://github.com/gatsbyjs/gatsby/commit/0558443))
+
 ## [2.5.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.5.3...gatsby-plugin-sharp@2.5.4) (2020-04-04)
 
 ### Bug Fixes