+++ packages/gatsby-transformer-sqip/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.7](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sqip@2.2.6...gatsby-transformer-sqip@2.2.7) (2020-04-24)
+
+**Note:** Version bump only for package gatsby-transformer-sqip
+
+## [2.2.6](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sqip@2.2.5...gatsby-transformer-sqip@2.2.6) (2020-04-17)
+
+### Bug Fixes
+
+- wrap ignore pattern in quotes ([#23176](https://github.com/gatsbyjs/gatsby/issues/23176)) ([7563db6](https://github.com/gatsbyjs/gatsby/commit/7563db6))
+
+## [2.2.5](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sqip@2.2.4...gatsby-transformer-sqip@2.2.5) (2020-04-16)
+
+**Note:** Version bump only for package gatsby-transformer-sqip
+
 ## [2.2.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sqip@2.2.3...gatsby-transformer-sqip@2.2.4) (2020-04-04)
 
 **Note:** Version bump only for package gatsby-transformer-sqip